Fix the 'active' counter
The 'active' counter showed enemies as active in missions like 'protect
the device'. Although that's technically correct (it showed that X
enemies are chacing the device), that's not what we wanted to display
here. We should now only count enemies which are aware of XCOMs presence
(by spotting them or whatever), so we only have the 'triggered' count
and not enemies which are wandering around shooting targets.