A 'WIP' snapshot of a spoiler free version
We moved away from the EventManager to utilize the VisualizationMgr
exclusively. This seems to work pretty well, although the
'OnVisualizationBlockComplete' handler implementation in this commit is
probably completly useless as it seems to be fine to just update on
every 'OnVisualizationIdle' event (as this seems to be the moment when
control is given back to the player).

The 'real time' feeling will go away this way but maybe there's another
way... (as GameStates come in out of order as they might take a long
time to visualize).