Fix a very rare lockup.
The loop basically ran away until the integer wrapped. This took a very
long time (several minutes on my machine) and fully locked the main game
thread. So it looked like a crash, but really was 'only' a lockup.
Unfortunately the state of the Mod was broken from there on so a
restart couldn't be circumvented.

Thank you Harddk for providing a savegame which let me easily debug
this!