Add debug switch
Previous build had (excessive) debug logging disabled as I feared it
would case a performance penalty. As we now have MCM, people can turn it
on when they need it. The code didn't got nicer by adding all the
if(debug) checks, but I was too lazy to create a helper (macro).