Make it WotC ready
After figuring out that the logic this mod used was completely useless
AND broke it for WotC, that part is now removed and replaced with a
simple VisualizationManager check (too bad I didn't knew that
earlier...).

When build with the classic SDK, the code does work for WotC as well,
even though the launcher thinks it might not (shows the triangle).